fixed blank_pixels taking forever, fixed random gen always giving same numbers

+++ include/visualizer/rectangle.h
@@ -7,7 +7,7 @@ namespace visualizer {
 
 class Rectangle : public Shape {
 public:
-  Rectangle(glm::vec2& corner, float width, float height, ci::ColorA& color);
+  Rectangle(glm::vec2& corner, int width, int height, ci::ColorA& color);
 
   void Draw() override;
   void Mutate() override;
@@ -15,14 +15,14 @@ public:
   glm::vec2 GetLocation() const override;
   void SetLocation(glm::vec2& new_corner) override;
 
-  float GetHeight() override;
-  float GetWidth() override;
+  int GetHeight() override;
+  int GetWidth() override;
   ci::ColorA GetColor() override;
 
 private:
   glm::vec2 top_left_corner_;
-  float height_;    // vertical side / y
-  float width_;     // horizontal side / x
+  int height_;    // vertical side / y
+  int width_;     // horizontal side / x
   ci::ColorA color_;
 
 };
