+++ include/visualizer/shape.h
@@ -14,14 +14,14 @@ public:
   virtual glm::vec2 GetLocation() const = 0;
   virtual void SetLocation(glm::vec2& new_loc) = 0;
 
-  virtual float GetWidth() = 0;
-  virtual float GetHeight() = 0;
+  virtual int GetWidth() = 0;
+  virtual int GetHeight() = 0;
   virtual ci::ColorA GetColor() = 0;
 
 private:
   glm::vec2 location_;
-  float width_;
-  float height_;
+  int width_;
+  int height_;
   ci::ColorA color_;
 
 };
