+++ src/visualizer/rectangle.cc
@@ -4,7 +4,7 @@
 namespace geometricprimitives {
 namespace visualizer {
 
-Rectangle::Rectangle(glm::vec2 &corner, float width, float height, ci::ColorA &color) : top_left_corner_(corner),
+Rectangle::Rectangle(glm::vec2 &corner, int width, int height, ci::ColorA &color) : top_left_corner_(corner),
                       width_(width), height_(height), color_(color) { }
 
 void Rectangle::Draw() {
@@ -19,8 +19,8 @@ void Rectangle::Mutate() {
 
 glm::vec2 Rectangle::GetLocation() const { return top_left_corner_; }
 void Rectangle::SetLocation(glm::vec2 &new_corner) { top_left_corner_ = new_corner; }
-float Rectangle::GetHeight() { return height_; }
-float Rectangle::GetWidth() { return width_; }
+int Rectangle::GetHeight() { return height_; }
+int Rectangle::GetWidth() { return width_; }
 ci::ColorA Rectangle::GetColor() { return color_; }
